Fixed isCompleted so that it uses the interface method
Fixed test cases so thart tests that test cryptogramPair are there own class
Fixed some method in cryptogram
