+++ tests/models/CryptogramTest.java
@@ -21,29 +21,6 @@ import pg13.models.CryptogramPair;
 		{
 			super(arg0);
 		}
-
-		public void testCryptogramPairClass()
-		{
-			CryptogramPair cp = new CryptogramPair('L', 'L');
-			assertNotNull(cp);
-		}
-		
-		public void testCryptogramPairClassCharacters()
-		{
-			assertNotNull(new CryptogramPair('L', 'L'));	
-			assertNotNull(new CryptogramPair('L', 'R'));	
-			assertNotNull(new CryptogramPair('\0', '\0'));
-			assertNotNull(new CryptogramPair('L', '\0'));
-			try
-			{
-				CryptogramPair cp1 = new CryptogramPair('!', '!');
-				fail();
-			}
-			catch(IllegalArgumentException iae)
-			{
-				
-			}
-		}
 		
 		public void testEmptyCryptogram()
 		{
